Rewrite the label widget.
Pyglet's layout class is very hard to work with, and the approach I took
before did not work well with attributes that can change on the fly.  So
I rewrote the label widget to create a new layout each time the draw
function is called.  This seems pretty robust, and it's not even that
expensive because all the vertex list would be redrawn even if I kept
the same layout.