Remove the Clickable class.
The purpose of the Clickable class was to separate the code for behaving
like a button from the code for rendering a button with text and images,
to make it easier to create custom buttons.

However, the interface between Clickable and its subclass(es) was fairly
complicated and brittle-feeling.  This made me think that Clickable
wouldn't actually be a very useful superclass and that it was making
Button more complicated than it needed to be.