Write a function for placing rects in grids.
I've decided that an object would be better than a function at managing
a grid.  For one thing, it reduces the amount of state other classes
need to manage in order to use the grid, and for another it allows
access to a rich set of getters.  But in case I change my mind, I wanted
to commit the more functional approach I took first.