Add a class for making grids of rects.
The purpose of this class was to move the job of making grids of
rectangles out of the Grid widget and into a class that could also be
used by artists like Background.  This new grid class ended up much more
powerful than the old Grid widget (and actually more similar to HVBox),
so as a bonus, this refactoring will ultimately make Grid widget will
more powerful itself.

This commit also add unit tests for the new grid class.  To help pytest
distinguish between the unit tests (which it can run) and the GUI demos
(which it can't), I renamed all the demos from test_* to demo_*.