Rewrite Button using the new Background widget.
While doing this, I decided that Button would be more flexible if it
actually had a different Background widget for each state (and hid all
but one at a time) rather than having one widget that it kept updating.

This required me to implement Widget.hide() and Widget.unhide().  It
think these methods turned out well.  The code is quite simple, it
allows widgets to be hidden and shown, and there's very little
performance penalty for a widget having hidden children.