Change how widgets are repacked.
Before, only changes that made a widget bigger would trigger a repack.
This was meant to limit how many widgets would be affected by a repack,
but it (not surprisingly) caused a lot of weird behavior when things got
smaller (e.g. removing a child from a container).  Now a widget will
propagate a repack if it changed size at all.