Rewrite the Button class using modular pieces.
I was frustrated that the Button class was so monolithic, so I broke it
down into modular pieces (specifically Clickable, Deck, and Rollover)
and rewrote it.  The Button class is just as functional as before, but
it's simpler because it doesn't need to worry about facilitating the
implementation of every conceivable button.  It just takes care of the
most common cases, and the more complicated stuff can be built up from
the same modular parts.