Elaborate on why resize() does what it does.
I tried to add a check resize() to only call do_resize_children() if the
size of the widget actually changed.  But this caused all sorts of
problems children widgets not updating when they should, so I undid the
change and added a comment explaining why things need to be the way they
are.