Refactor padding.
I realized that it would be important to support the ability to have
different paddings on each side of a widget.  However, I ran into a
number of problems trying to implement this from the perspective of
container widgets.  First, I didn't want to add 7 padding arguments to
each method that added a widget to a container.  Second, I couldn't find
a good way to distinguish paddings that applied to the whole container
and paddings that applied only to certain children.

Instead, it occurred to me that all padding could actually be handled in
the base Widget class.  Previously Widget just had a rect that was the
shape given to it by its parent.  I tweaked this scheme such that Widget
is given a rect by its parent, but then calculates a padded rect for
it's own content.  This solved all the issues I was having very
parsimoniously.

The next steps will be to make the same change for the placement
attribute (which I want to rename to 'alignment'), to add class-level
default_padding* attributes, and to extend the padding system with an
inter-cell padding for Grid and HVBox containers.