Use generators to cycle through tests.
I found that I was writing a lot of tests that cycled through a number
of different configurations as the user pressed space.  This took a fair
amount of boilerplate, until I realized that I could write the tests as
a single generator and write a decorator to handle the logic of moving
between states.  This also made it easy to label active configuration,
so I added that feature as well.