Make alignment (placement) a Widget attribute.
Previously alignment was called placement and was handled exclusively by
container widgets.  However, it became possible for widgets to handle
this property themselves when I gave Widget the concept of an assigned
rect and a content rect (part of the effort to refactor padding).  This
approach turned out to meaningfully simplify the code and to increase
the flexibility you have in placing, composing, and configuring widgets.

This commit also includes some unrelated house-keeping:

- Fix padding and alignment issues in all of the demos.

- Use the new interactive_test framework to make a number of the demos
  more complete and rigorous.

- Add "del" methods to the Label class.

- Add Color.from_str() and Color.from_anything().