Add a function to debug widget placement problems.
Sometimes I just want to see the various rectangles associated with a
widget, to help understand why it's not showing up where I think it
should.  So I added a method to Widget to do this.  I think this new
method is similar in spirit to the method I already had to help figure
out why a widget wasn't drawing.