Centralize the logic for applying alignments.
Make an align() function in the drawing module that takes a string or
function and uses it to align two rectangles.  Previously the calling
code was responsible for first casting the string to a function, then
calling that function separately.  This commit better abstracts what an
alignment really is and simplifies the classes that use alignments.