Make Grid.find_cell_under_mouse() agree with Widget.is_under_mouse()
There was a 1 px mismatch between these two functions.  In combination
with a bug in the way spurious mouse events were being detected, this
was causing on_mouse_enter and on_mouse_leave events to be missed in
certain circumstances.

Even though fixing the bug in the spurious event handler fixes this
problem, I thought it would be prudent to have these two methods agree.