Only handle spurious events in the root widget.
The root widget is the only widget that really needs to handle spurious
events, because it knows enough to filter them out and no other widget
will see them after that.  But I had the code in Widget, and it was
causing bugs by reporting spurious events for widgets that reimplement
do_find_children_under_mouse() in a way that doesn't perfectly agree
with Widget.is_under_mouse().

The specific problem was a 1 px discrepancy between Grid/HVBox and
Widget.  I fixed that discrepancy, but I also moved the spurious event
handling code from Widget to Root so that do_find_children_under_mouse()
doesn't need to be in pixel-perfect agreement with is_under_mouse().