Handle on_mouse_leave events more robustly.
When a parent gets an on_mouse_leave event, it will now propagate the
event to every child that isn't grabbing the mouse.  Originally it would
propagate the event to every child.  When I added mouse grabbing, I had
to exclude the child grabbing the mouse, but I did this by reusing
existing code that also checked to coordinates of the child.  This ended
up creating some corner cases (maybe just with the Mover widget) where a
child could get stuck in the "over" mouse state when the mouse left its
parent but not it.