Don't repack if the window doesn't change size.
Pyglet sometimes emits Window.on_resize events even when the window
hasn't actually changed size (e.g. when it's exposed, I think).  The GUI
normally repacks in response to Window.on_resize events, but this commit
adds logic to skip the repack unless the window actually changed size.

I think this is mostly did this because it's useful to not have extra
repacks when you're trying to debug something.  The actual performance
improvement is surely negligible.