Allow widgets to fill scroll boxes.
The Mover/ScrollPane code was written such that their children (i.e.
widgets that would be scrolled or translated) were always made as small
as possible.  That behavior makes sense in any dimensions being scrolled
(since those dimensions don't have maximum sizes)), but not in any
dimensions being shown in full.

Changing this behavior meant rewriting Mover.do_resize_children() to
assign the maximum size in the dimensions that ask for it, and rewriting
parts of ScrollPane to take advantage of this.