Don't allow text to fill all the width available to it.
The problem is that text claims a certain width and height.  If it then
gets more width than it asked for, and re-wraps itself to fill all of
it, you'll end up with bunch of unused space below the text.

The real solution would be to have text claim its width first and then
its height based on that.  But this seems pretty difficult, and may
require a significant refactoring of the repacking code.  In the mean
time, I'm just not allowing text to fill extra width allocated to it.
This is somewhat limiting (although  not as bad for game GUIs as general
purpose GUIs, I think), but at least it's not surprising or wrong.