Make Widget responsible for `on_click` and related events.
Previously I had a Clickable class that managed all the events for
widgets that could be clicked on.  But as I was writing a demo on making
a checkbox with a clickable label, it occurred to me that it was
confusing to have Widget handle some of the mouse-related events
(on_mouse_press, etc) but not all of them.  So I decided to merge
Clickable into Widget.