Add an option to prevent the GUI from clearing the window.
Clearing the screen actually seems like a pretty disruptive thing to be
doing by default.  This commit doesn't change the default behavior
(which is to clear the screen each time the gui is drawn) but I think I
need to think more about this.

Maybe it would be better to have users subclass `Gui` and overwrite
on_draw() to redraw to redraw whatever needs to be redrawn after the
clear().  I'm not sure.

In the meantime, this is probably a good option to have.  Fixes #22, I
think.