Generalize th Button interface.
Rather than having hard coded "button" and "image" layers, I now have a
"foreground" layer that can be any kind of widget.  Note that this is a
backwards-incompatible change.  I also removed the internal stack that
was being used to manage the layers, because it's pretty easy to manage
two layers, and Button is common enough I think it's worth minimizing
the depth of the widget hierarchy.

This commit also includes updates to the documentation (since there are
a lot of the buttons ins the docs) and fixes a scroll bar crash that I
found while going through the docs.

The scroll bar change is actually something of a refactor: I changed the
GripWrapper class to a GripMover class, and calculate the size of the
grip in do_resize_children() rather than do_claim().  This is more
robust (because the sizes I need are not guaranteed to exist when claims
are being made) and it removes a level from the widget hierarchy, so its
all good!