Allow alignment error-checking to be bypassed.
I recently added a bunch of stringent error-checks to the alignment
functions in response to a nasty bug caused by such a function
misbehaving.  In particular, I required that the child rectangle remain
smaller than the parent rectangle.

This requirement turned out to be problematic in the context of the
ScrollPane, which was using the alignment functions to positiong the
*pane* inside the *child* (opposite of the usual child inside parent).
In the unusual-but-possible case that the child was smaller than the
scroll pane, this would trigger the above error-check, even though there
was really no problem in this case.

I handled this by adding an argument to the alignment function allowing
this particular error-cehck to be skipped.